requirements.txt
pillow
KindleComicConverter
cloudscraper
